+++ src/deesseia/__init__.py
@@ -4,7 +4,7 @@ Deesseia - Goddess of Data Science.
 A comprehensive data science library that streamlines the entire data science workflow from data loading to model deployment.
 """
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 
 from deesseia.core.base.cleaner import BaseCleaner
 from deesseia.core.base.inspector import BaseDataInspector
@@ -14,6 +14,8 @@ from deesseia.core.cleaner import Cleaner
 from deesseia.core.inspector import DataInspector
 from deesseia.core.loader import DataLoader
 from deesseia.core.validator import Validator
+from deesseia.preprocess.encoder import Encoder
+from deesseia.preprocess.scaler import Scaler
 from deesseia.utils.fake_data import FakeDataGenerator
 
 __all__ = [
@@ -24,6 +26,8 @@ __all__ = [
     "Cleaner",
     "DataInspector",
     "DataLoader",
+    "Encoder",
     "FakeDataGenerator",
+    "Scaler",
     "Validator",
 ]